% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiosampleApi.r
\name{BiosampleApi}
\alias{BiosampleApi}
\title{Biosample operations}
\description{
swagger.Biosample
}
\section{Methods}{

\describe{

biosample_by_accession_biosample_samples_accession_get Biosample By Accession


biosamples_biosample_samples_get Biosamples


mapping_biosample_fields_entity_get Mapping

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BiosampleApi$new()}}
\item \href{#method-biosample_by_accession_biosample_samples_accession_get}{\code{BiosampleApi$biosample_by_accession_biosample_samples_accession_get()}}
\item \href{#method-biosamples_biosample_samples_get}{\code{BiosampleApi$biosamples_biosample_samples_get()}}
\item \href{#method-mapping_biosample_fields_entity_get}{\code{BiosampleApi$mapping_biosample_fields_entity_get()}}
\item \href{#method-clone}{\code{BiosampleApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiosampleApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-biosample_by_accession_biosample_samples_accession_get"></a>}}
\if{latex}{\out{\hypertarget{method-biosample_by_accession_biosample_samples_accession_get}{}}}
\subsection{Method \code{biosample_by_accession_biosample_samples_accession_get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiosampleApi$biosample_by_accession_biosample_samples_accession_get(
  accession,
  include_fields,
  exclude_fields,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-biosamples_biosample_samples_get"></a>}}
\if{latex}{\out{\hypertarget{method-biosamples_biosample_samples_get}{}}}
\subsection{Method \code{biosamples_biosample_samples_get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiosampleApi$biosamples_biosample_samples_get(
  q,
  size,
  cursor,
  facet_size,
  include_fields,
  exclude_fields,
  facets,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mapping_biosample_fields_entity_get"></a>}}
\if{latex}{\out{\hypertarget{method-mapping_biosample_fields_entity_get}{}}}
\subsection{Method \code{mapping_biosample_fields_entity_get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiosampleApi$mapping_biosample_fields_entity_get(entity, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BiosampleApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
