% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultApi.r
\name{DefaultApi}
\alias{DefaultApi}
\title{Default operations}
\description{
swagger.Default
}
\section{Methods}{

\describe{

facets_by_index_facets_entity_get Facets By Index

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DefaultApi$new()}}
\item \href{#method-facets_by_index_facets_entity_get}{\code{DefaultApi$facets_by_index_facets_entity_get()}}
\item \href{#method-clone}{\code{DefaultApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-facets_by_index_facets_entity_get"></a>}}
\if{latex}{\out{\hypertarget{method-facets_by_index_facets_entity_get}{}}}
\subsection{Method \code{facets_by_index_facets_entity_get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$facets_by_index_facets_entity_get(entity, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
