% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_connection.R
\name{get_omicidx_duckdb_location}
\alias{get_omicidx_duckdb_location}
\title{Get the location of the OmicIDX DuckDB database file.}
\usage{
get_omicidx_duckdb_location()
}
\value{
A string representing the URL of the OmicIDX DuckDB database file.
}
\description{
This function returns the URL of the OmicIDX DuckDB database file.
You can use this URL to access the database directly or for other purposes.
}
\examples{
url <- get_omicidx_duckdb_location()
print(url)

}
