% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{Omicidx}
\alias{Omicidx}
\title{R6 Class for Omicidx Database Interaction}
\description{
The `Omicidx` R6 class provides an interface for interacting with the omicidx database.
}
\examples{
# # Create a new Omicidx object
omicidx <- Omicidx$new()
# # List all table names in the database
omicidx$table_names()
# Get a dplyr table object for a specific table
sra_studies_tbl <- omicidx$tbl("sra_studies")
# Perform a dplyr query
result <- sra_studies_tbl |>
  filter(study_type == "RNA-Seq") |>
  head(10) |>
  collect()
print(result)

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{connection}}{A DBI connection object to the omicidx database.}

\item{\code{.dplyr_tbls}}{A list to cache dplyr table objects.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Omicidx-new}{\code{Omicidx$new()}}
\item \href{#method-Omicidx-sql}{\code{Omicidx$sql()}}
\item \href{#method-Omicidx-table_names}{\code{Omicidx$table_names()}}
\item \href{#method-Omicidx-tbl}{\code{Omicidx$tbl()}}
\item \href{#method-Omicidx-clone}{\code{Omicidx$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Omicidx-new"></a>}}
\if{latex}{\out{\hypertarget{method-Omicidx-new}{}}}
\subsection{Method \code{new()}}{
Create a new `Omicidx` object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omicidx$new(con = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{con}}{An optional DBI connection object. If NULL, a new connection will be created.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Omicidx` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Omicidx-sql"></a>}}
\if{latex}{\out{\hypertarget{method-Omicidx-sql}{}}}
\subsection{Method \code{sql()}}{
Execute a SQL query on the omicidx database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omicidx$sql(query)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{query}}{A string containing the SQL query to execute.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame containing the results of the query.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Omicidx-table_names"></a>}}
\if{latex}{\out{\hypertarget{method-Omicidx-table_names}{}}}
\subsection{Method \code{table_names()}}{
List all table names in the omicidx database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omicidx$table_names()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data frame containing the names of all tables in the database.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Omicidx-tbl"></a>}}
\if{latex}{\out{\hypertarget{method-Omicidx-tbl}{}}}
\subsection{Method \code{tbl()}}{
Get a dplyr table object for a specified table in the omicidx database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omicidx$tbl(table_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table_name}}{A string specifying the name of the table.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This method retrieves a dplyr table object for the specified table name.
If the table has been previously accessed, it will return the cached version.
If the table_name does not exist, an error will be raised.
}

\subsection{Returns}{
A dplyr table object for the specified table.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Omicidx-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Omicidx-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omicidx$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
