% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_introspection.R
\name{omicidx_db_tables}
\alias{omicidx_db_tables}
\title{List Database Tables in the omicidx database}
\usage{
omicidx_db_tables(con, schema = NULL)
}
\arguments{
\item{con}{A DBI connection object to the omicidx database.}
}
\value{
A data frame containing the names of all tables in the database.
}
\description{
This function retrieves a list of all tables in the connected omicidx database.
}
\examples{
\dontrun{
con <- omicidx_duckdb_connection()
tables <- omicidx_db_tables(con)
print(tables)
}
}
