% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_connection.R
\name{omicidx_duckdb_connection}
\alias{omicidx_duckdb_connection}
\title{Establish a read-only connection to the OmicIDX DuckDB database.}
\usage{
omicidx_duckdb_connection()
}
\value{
A DBI connection object to the OmicIDX DuckDB database.
}
\description{
The omicidx DuckDB database is hosted remotely and accessed via the httpfs extension.
This function sets up a connection to the database for querying. The connection
is a normal DuckDB connection, but the database file is read-only.
}
\examples{
\dontrun{
con <- omicidx_duckdb_connection()
DBI::dbListTables(con)
# Example query
res <- DBI::dbGetQuery(con, "SELECT * FROM sra_runs LIMIT 10")
print(res)
}

}
