% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRAApi.r
\name{SRAApi}
\alias{SRAApi}
\title{SRA operations}
\description{
swagger.SRA
}
\section{Methods}{

\describe{

experiment_by_accession_sra_experiments_accession_get Experiment By Accession


experiments_for_sample_sra_samples_accession_experiments_get Experiments For Sample


experiments_for_study_sra_studies_accession_experiments_get Experiments For Study


experiments_sra_experiments_get Experiments


mapping_sra_fields_entity_get Mapping


run_by_accession_sra_runs_accession_get Run By Accession


runs_for_experiment_sra_experiments_accession_runs_get Runs For Experiment


runs_for_sample_sra_samples_accession_runs_get Runs For Sample


runs_for_study_sra_studies_accession_runs_get Runs For Study


runs_sra_runs_get Runs


sample_by_accession_sra_samples_accession_get Sample By Accession


samples_for_study_sra_studies_accession_samples_get Samples For Study


samples_sra_samples_get Samples


studies_sra_studies_get Studies


study_by_accession_sra_studies_accession_get Study By Accession

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SRAApi$new()}}
\item \href{#method-experiment_by_accession_sra_experiments_accession_get}{\code{SRAApi$experiment_by_accession_sra_experiments_accession_get()}}
\item \href{#method-experiments_for_sample_sra_samples_accession_experiments_get}{\code{SRAApi$experiments_for_sample_sra_samples_accession_experiments_get()}}
\item \href{#method-experiments_for_study_sra_studies_accession_experiments_get}{\code{SRAApi$experiments_for_study_sra_studies_accession_experiments_get()}}
\item \href{#method-experiments_sra_experiments_get}{\code{SRAApi$experiments_sra_experiments_get()}}
\item \href{#method-mapping_sra_fields_entity_get}{\code{SRAApi$mapping_sra_fields_entity_get()}}
\item \href{#method-run_by_accession_sra_runs_accession_get}{\code{SRAApi$run_by_accession_sra_runs_accession_get()}}
\item \href{#method-runs_for_experiment_sra_experiments_accession_runs_get}{\code{SRAApi$runs_for_experiment_sra_experiments_accession_runs_get()}}
\item \href{#method-runs_for_sample_sra_samples_accession_runs_get}{\code{SRAApi$runs_for_sample_sra_samples_accession_runs_get()}}
\item \href{#method-runs_for_study_sra_studies_accession_runs_get}{\code{SRAApi$runs_for_study_sra_studies_accession_runs_get()}}
\item \href{#method-runs_sra_runs_get}{\code{SRAApi$runs_sra_runs_get()}}
\item \href{#method-sample_by_accession_sra_samples_accession_get}{\code{SRAApi$sample_by_accession_sra_samples_accession_get()}}
\item \href{#method-samples_for_study_sra_studies_accession_samples_get}{\code{SRAApi$samples_for_study_sra_studies_accession_samples_get()}}
\item \href{#method-samples_sra_samples_get}{\code{SRAApi$samples_sra_samples_get()}}
\item \href{#method-studies_sra_studies_get}{\code{SRAApi$studies_sra_studies_get()}}
\item \href{#method-study_by_accession_sra_studies_accession_get}{\code{SRAApi$study_by_accession_sra_studies_accession_get()}}
\item \href{#method-clone}{\code{SRAApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SRAApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-experiment_by_accession_sra_experiments_accession_get"></a>}}
\if{latex}{\out{\hypertarget{method-experiment_by_accession_sra_experiments_accession_get}{}}}
\subsection{Method \code{experiment_by_accession_sra_experiments_accession_get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SRAApi$experiment_by_accession_sra_experiments_accession_get(
  accession,
  include_fields,
  exclude_fields,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-experiments_for_sample_sra_samples_accession_experiments_get"></a>}}
\if{latex}{\out{\hypertarget{method-experiments_for_sample_sra_samples_accession_experiments_get}{}}}
\subsection{Method \code{experiments_for_sample_sra_samples_accession_experiments_get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SRAApi$experiments_for_sample_sra_samples_accession_experiments_get(
  accession,
  include_fields,
  exclude_fields,
  size,
  cursor,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-experiments_for_study_sra_studies_accession_experiments_get"></a>}}
\if{latex}{\out{\hypertarget{method-experiments_for_study_sra_studies_accession_experiments_get}{}}}
\subsection{Method \code{experiments_for_study_sra_studies_accession_experiments_get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SRAApi$experiments_for_study_sra_studies_accession_experiments_get(
  accession,
  include_fields,
  exclude_fields,
  size,
  cursor,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-experiments_sra_experiments_get"></a>}}
\if{latex}{\out{\hypertarget{method-experiments_sra_experiments_get}{}}}
\subsection{Method \code{experiments_sra_experiments_get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SRAApi$experiments_sra_experiments_get(
  q,
  size,
  cursor,
  facet_size,
  include_fields,
  exclude_fields,
  facets,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mapping_sra_fields_entity_get"></a>}}
\if{latex}{\out{\hypertarget{method-mapping_sra_fields_entity_get}{}}}
\subsection{Method \code{mapping_sra_fields_entity_get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SRAApi$mapping_sra_fields_entity_get(entity, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_by_accession_sra_runs_accession_get"></a>}}
\if{latex}{\out{\hypertarget{method-run_by_accession_sra_runs_accession_get}{}}}
\subsection{Method \code{run_by_accession_sra_runs_accession_get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SRAApi$run_by_accession_sra_runs_accession_get(
  accession,
  include_fields,
  exclude_fields,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-runs_for_experiment_sra_experiments_accession_runs_get"></a>}}
\if{latex}{\out{\hypertarget{method-runs_for_experiment_sra_experiments_accession_runs_get}{}}}
\subsection{Method \code{runs_for_experiment_sra_experiments_accession_runs_get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SRAApi$runs_for_experiment_sra_experiments_accession_runs_get(
  accession,
  include_fields,
  exclude_fields,
  size,
  cursor,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-runs_for_sample_sra_samples_accession_runs_get"></a>}}
\if{latex}{\out{\hypertarget{method-runs_for_sample_sra_samples_accession_runs_get}{}}}
\subsection{Method \code{runs_for_sample_sra_samples_accession_runs_get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SRAApi$runs_for_sample_sra_samples_accession_runs_get(
  accession,
  include_fields,
  exclude_fields,
  size,
  cursor,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-runs_for_study_sra_studies_accession_runs_get"></a>}}
\if{latex}{\out{\hypertarget{method-runs_for_study_sra_studies_accession_runs_get}{}}}
\subsection{Method \code{runs_for_study_sra_studies_accession_runs_get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SRAApi$runs_for_study_sra_studies_accession_runs_get(
  accession,
  include_fields,
  exclude_fields,
  size,
  cursor,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-runs_sra_runs_get"></a>}}
\if{latex}{\out{\hypertarget{method-runs_sra_runs_get}{}}}
\subsection{Method \code{runs_sra_runs_get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SRAApi$runs_sra_runs_get(
  q,
  size,
  cursor,
  facet_size,
  include_fields,
  exclude_fields,
  facets,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample_by_accession_sra_samples_accession_get"></a>}}
\if{latex}{\out{\hypertarget{method-sample_by_accession_sra_samples_accession_get}{}}}
\subsection{Method \code{sample_by_accession_sra_samples_accession_get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SRAApi$sample_by_accession_sra_samples_accession_get(
  accession,
  include_fields,
  exclude_fields,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-samples_for_study_sra_studies_accession_samples_get"></a>}}
\if{latex}{\out{\hypertarget{method-samples_for_study_sra_studies_accession_samples_get}{}}}
\subsection{Method \code{samples_for_study_sra_studies_accession_samples_get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SRAApi$samples_for_study_sra_studies_accession_samples_get(
  accession,
  include_fields,
  exclude_fields,
  size,
  cursor,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-samples_sra_samples_get"></a>}}
\if{latex}{\out{\hypertarget{method-samples_sra_samples_get}{}}}
\subsection{Method \code{samples_sra_samples_get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SRAApi$samples_sra_samples_get(
  q,
  size,
  cursor,
  facet_size,
  include_fields,
  exclude_fields,
  facets,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-studies_sra_studies_get"></a>}}
\if{latex}{\out{\hypertarget{method-studies_sra_studies_get}{}}}
\subsection{Method \code{studies_sra_studies_get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SRAApi$studies_sra_studies_get(
  q,
  size,
  cursor,
  facet_size,
  include_fields,
  exclude_fields,
  facets,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-study_by_accession_sra_studies_accession_get"></a>}}
\if{latex}{\out{\hypertarget{method-study_by_accession_sra_studies_accession_get}{}}}
\subsection{Method \code{study_by_accession_sra_studies_accession_get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SRAApi$study_by_accession_sra_studies_accession_get(
  accession,
  include_fields,
  exclude_fields,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SRAApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
